# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: ALKABRAIN — AI Router with auth, billing, and self-deployment.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /webhooks/razorpay:
    post:
      operationId: razorpayWebhook
      tags: [webhooks]
      summary: Razorpay webhook receiver (HMAC verified)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RazorpayWebhookEvent"
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
        "400":
          description: Invalid signature
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
...
  /deploy/push:
    post:
      operationId: triggerDeploy
      tags: [deploy]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeployRequest" }
      responses:
        "200":
          description: Deployment started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeployJob" }
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    RazorpayWebhookEvent:
      type: object
      additionalProperties: true
      properties:
        event: { type: string }
    DeployRequest:
      type: object
      properties:
        commitMessage: { type: string }
...
    DeployJob:
      type: object
      required: [id, startedAt, state, spaceRepo]
      properties:
        id: { type: string }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time, nullable: true }
        state:
          type: string
          enum: [pending, pushing, building, success, failed]
        spaceRepo: { type: string }
        commitMessage: { type: string, nullable: true }
        commitSha: { type: string, nullable: true }
        message: { type: string, nullable: true }
        logs: { type: string, nullable: true }
...